% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{CalcPhyDispStats}
\alias{CalcPhyDispStats}
\title{CalcPhyDispStats}
\usage{
CalcPhyDispStats(regional.tree, local.tree, regional.traits, local.traits)
}
\arguments{
\item{regional.tree:}{regional community phylogenetic tree, of class "phylo"}

\item{local.tree:}{local community phylogenetic tree, of class "phylo"}

\item{regional.traits:}{trait values for each species in regional community, vector or list}

\item{local.traits:}{trait values for each species in local community, vector or list}
}
\value{
vector of output from ses.mpd and ses.mndt
}
\description{
CalcPhyDispStats
}
\examples{
: 
}

