% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{CalcSummaryStats}
\alias{CalcSummaryStats}
\title{CalcSummaryStats}
\usage{
CalcSummaryStats(regional.tree, local.tree, regional.traits, local.traits)
}
\arguments{
\item{regional.tree:}{regional community phylogenetic tree, of class "phylo"}

\item{local.tree:}{local community phylogenetic tree, of class "phylo"}

\item{regional.traits:}{trait values for each species in regional community, vector or list}

\item{local.traits:}{trait values for each species in local community, vector or list}
}
\value{
vector of summary statistics
}
\description{
CalcSummaryStats
}
\examples{
: 
}

