% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{SimCommunityAssembly}
\alias{SimCommunityAssembly}
\title{SimCommunityAssembly}
\usage{
SimCommunityAssembly(sims, N, local, traitsim, comsim, disable.bar = FALSE,
  output.sum.stats = TRUE, output.phydisp.stats = FALSE, lambda = c(0.05,
  2), eps = c(0.2, 0.8), sig2 = c(1, 10), alpha = c(0.01, 0.2),
  tau = c(5, 50))
}
\arguments{
\item{sims:}{The number of simulations; sims}

\item{N:}{The size of the regional communities, can be a single number or a prior range}

\item{local:}{size of the local communities, can be a single number, a proportion of the regional pool (btw 0 and 1), or a prior range of proportions}

\item{traitsim:}{either 'BM' or 'OU'}

\item{comsim:}{either "netural", "filtering", "competition"}

\item{lambda:}{speciation rate, can be single number or prior range, default is a unifrom prior between 0.5 and 2.0}

\item{eps:}{the extinction fraction, extinction (mu) = lambda * eps, default is a uniform prior between 0.2 and 0.8}

\item{sig2:}{rate of character change, can be single value or prior range, defualt is a unifom prior range between 1.0 and 10.0}

\item{alpha:}{pull to trait optimum, defualt is a uniform prior range between 0.01 and 0.2}

\item{tau:}{stregnth of the community assembly model, default is a uniform prior range between 5.0 and 5.0}
}
\value{
output is a list of two matrices, one containing all parameter values for each simulation, and the other containing all 
        summary statistics. In both matrices, each row corresponds to one simulation
}
\description{
SimCommunityAssembly
}

